% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysis_means}
\alias{analysis_means}
\title{FUNCTION: analysis_means}
\usage{
analysis_means(data)
}
\arguments{
\item{data}{A data frame of read_analysis}
}
\value{
A Data Frame of averages
}
\description{
Averages of test generation timing and coverage
}
