% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysis_medians}
\alias{analysis_medians}
\title{FUNCTION: analysis_medians}
\usage{
analysis_medians(data)
}
\arguments{
\item{data}{A data frame of read_analysis}
}
\value{
A Data Frame of medians
}
\description{
Medians of test generation timing and coverage
}
