% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{comparing_sig}
\alias{comparing_sig}
\title{FUNCTION: dominoR::comparing_sig}
\usage{
comparing_sig(sample1, sample2, effect, result)
}
\value{
latex of cell
}
\description{
FUNCTION: dominoR::comparing_sig
}
