% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{comparing_sig_timing}
\alias{comparing_sig_timing}
\title{FUNCTION: comparing_sig_timing}
\usage{
comparing_sig_timing(sample1, sample2, effect, result)
}
\value{
latex of cell
}
\description{
FUNCTION: comparing_sig_timing
}
