% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{concentro_table_combaining}
\alias{concentro_table_combaining}
\title{FUNCTION: concentro_table_combaining}
\usage{
concentro_table_combaining(ana, mut, rtrn = "tex")
}
\arguments{
\item{ana}{Data frame of analysis}

\item{mut}{Data frame of mutants}
}
\value{
A data frame of check constraints only schemas or table LaTeX
}
\description{
Timing, Coverage and mutation scores combained.
}
