% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{demo_timing_plot}
\alias{demo_timing_plot}
\title{FUNCTION: demo_timing_plot}
\usage{
demo_timing_plot(analysis)
}
\arguments{
\item{analysis}{A data frame of read_analysis}
}
\value{
A plot
}
\description{
Timing Plot for ICST slides
}
