% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{domino_table_combined}
\alias{domino_table_combined}
\title{FUNCTION: domino_table_combined}
\usage{
domino_table_combined(ana, mut, rtrn = "tex", mm = "median")
}
\arguments{
\item{ana}{Data frame of analysis}

\item{mut}{Data frame of mutants}
}
\value{
A data frame of check constraints only schemas or table LaTeX
}
\description{
Timing, Coverage and mutation scores combained.
}
