% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{get_check_schemas_mutants}
\alias{get_check_schemas_mutants}
\title{FUNCTION: get_check_schemas_mutants}
\usage{
get_check_schemas_mutants(d)
}
\arguments{
\item{d}{Data frame of analysis}
}
\value{
A data frame of check constraints only schemas
}
\description{
Filtering analysis data to get ony check constraints schemas.
}
