% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{ordering_mutants_per_operator_concentro}
\alias{ordering_mutants_per_operator_concentro}
\title{FUNCTION: ordering_mutants_per_operator}
\usage{
ordering_mutants_per_operator_concentro(d)
}
\arguments{
\item{d}{Data frame of mutants}
}
\value{
A data frame of ordred mutants and grouped by runs and mutation score per run per operator
}
\description{
It generates an ordered data frame of mutants (normal type) grouped by each run per operator and its mutation score.
}
