% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{ordering_mutants_per_schema_others}
\alias{ordering_mutants_per_schema_others}
\title{FUNCTION: ordering_mutants_per_schema_others}
\usage{
ordering_mutants_per_schema_others(d)
}
\arguments{
\item{d}{Data frame of mutants}
}
\value{
A data frame of ordred mutants and grouped by runs and mutation score per run per schema
}
\description{
It generates an ordered data frame of mutants (normal type) grouped by each run per schema and its mutation score.
}
