% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{read_failedtest_with_filtering}
\alias{read_failedtest_with_filtering}
\title{FUNCTION: read_failedtest_with_filtering}
\usage{
read_failedtest_with_filtering(dt, type = "ana")
}
\value{
A Data Frame of analysis
}
\description{
Read the data file that contains the "analysis" data. This is the data containing all test generation times, coverages,
evaluations etc. It is refered in SchemaAnalyst github repo as 'newmutationanalysis.dat'. And it allow us to compare
test generation timing and coverages results.
}
