% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{slides_coverage_hybrid_plot}
\alias{slides_coverage_hybrid_plot}
\title{FUNCTION: slides_coverage_hybrid_plot}
\usage{
slides_coverage_hybrid_plot(analysis)
}
\arguments{
\item{analysis}{A data frame of read_analysis}
}
\value{
A plot
}
\description{
Coverage for DominoAVM Plot for ICST slides
}
