% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{slides_mutation_plot}
\alias{slides_mutation_plot}
\title{FUNCTION: slides_mutation_plot}
\usage{
slides_mutation_plot(mutation)
}
\arguments{
\item{mutation}{A data frame of read_analysis}
}
\value{
A plot
}
\description{
Mutation Plot for ICST slides
}
