% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{slides_timing_hybrid_plot}
\alias{slides_timing_hybrid_plot}
\title{FUNCTION: slides_timing_hybrid_plot}
\usage{
slides_timing_hybrid_plot(analysis)
}
\arguments{
\item{analysis}{A data frame of read_analysis}
}
\value{
A plot
}
\description{
Timing Plot for ICST slides
}
