% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{table_generator_coverage_domino}
\alias{table_generator_coverage_domino}
\title{FUNCTION: table_generator_coverage_domino}
\usage{
table_generator_coverage_domino(d, rtrn = "tex", m = "median")
}
\arguments{
\item{d}{Data frame of analysis}

\item{rtrn}{Latex (tex) or a data frame (data)}

\item{m}{Results shown as median or mean}
}
\value{
A A12 effect size and U-test of coverages compared pair wise
}
\description{
Generates a latex table or a data frame for coverage table with effect size and U test.
Only for domino AVM and Random
}
