% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{table_generator_mutation_score_concentro}
\alias{table_generator_mutation_score_concentro}
\title{FUNCTION: table_generator_mutation_score_concentro}
\usage{
table_generator_mutation_score_concentro(d, rtrn = "tex", m = "median")
}
\arguments{
\item{d}{Data frame of mutants}

\item{rtrn}{Latex (tex) or a data frame (data)}

\item{m}{Results shown as median or mean}
}
\value{
A A12 effect size and U-test of mutation score per schema compared pair wise
}
\description{
Generates a latex table or data frame for mutation score per schema table with effect size and U test.
Only for Concentro AVM and Random
}
